Try comparing records via marc.to_s hash values
To increase performance: after detecting overlays, compare marc.to_s
hash values for the incoming and existing records. Only if the hash
values do not match do we need to continue on to more elaborate
(e.g. normalized, etc.) matching. The 005s are dropped before hashing
because they will always differ in WCM records. Hence this quicker
matching routing is skipped if 005s are not listed as omitted from
comparison in the specs.